trigger login event after session created
This plugin works well when we use a CAS login page, but if we have a system with a custom login mechanism,  and we cannot change it, we have to use CAS rest api  to login to a service.
When using this method, we open a url like http://example.com/?ticket=xxxxx to login. If we do not subscribe core.session_create_after, we have to open the url twice to login.
I'm not familiar with php and phpbb, but after two days' work, I found a nice solution. I don't know whether it's correct or I have missed something. But it works for me.